pyproject.toml
[tool.poetry]
name = "fastapi-production-ready"
version = "0.1.0"
description = ""
authors = ["pavelborisofff <pavel.borisofff@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_production_ready"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
httpx = "^0.24.0"
mypy = "^1.2.0"
pylint = "^2.17.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
